% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_district_literacy_df.R,
%   R/new_literacy_data.R
\name{new_literacy_data}
\alias{new_literacy_data}
\title{Literacy Dataframe Class Constructor}
\usage{
new_literacy_data(
  name_of_region = "Ashanti",
  name_of_district,
  census_year = 2010,
  total_literate_population,
  total_not_literate_population,
  number_literate_males,
  number_literate_females,
  number_not_literate_males,
  number_not_literate_females
)

new_literacy_data(
  name_of_region = "Ashanti",
  name_of_district,
  census_year = 2010,
  total_literate_population,
  total_not_literate_population,
  number_literate_males,
  number_literate_females,
  number_not_literate_males,
  number_not_literate_females
)
}
\arguments{
\item{name_of_region}{A character string specifying the region name.
Defaults to "Ashanti".}

\item{name_of_district}{A character string specifying the name of the district.}

\item{census_year}{Year in which the census data was collected.}

\item{total_literate_population}{An integer representing the total literate
population in the district.}

\item{total_not_literate_population}{An integer representing the total not
literate population in the district.}

\item{number_literate_males}{An integer representing the number of literate
males in the district.}

\item{number_literate_females}{An integer representing the number of literate
females in the district.}

\item{number_not_literate_males}{An integer representing the number of males
in the district who cannot read.}

\item{number_not_literate_females}{An integer representing the number of
females in the district who cannot read.}
}
\value{
An S3 object of class "literacy_data".

An S3 object of class "literacy_data".
}
\description{
Creates an S3 object of class \code{literacy_dataframe} containing district
literacy information.

Creates an S3 object of class "literacy_dataframe" containing district literacy
information.
}
\examples{
# Create literacy data for the Kumasi Metropolitan district
kumasi_literacy <- new_literacy_data(
  name_of_district = "Kumasi Metropolitan",
  census_year = 2010,
  total_literate_population = 800000,
  total_not_literate_population = 200000,
  number_literate_males = 450000,
  number_literate_females = 350000,
  number_not_literate_males = 100000,
  number_not_literate_females = 100000
)

# Create literacy data for the Kumasi Metropolitan district
kumasi_literacy <- new_literacy_data(
  name_of_district = "Kumasi Metropolitan",
  census_year = 2010,
  total_literate_population = 800000,
  total_not_literate_population = 200000,
  number_literate_males = 450000,
  number_literate_females = 350000,
  number_not_literate_males = 100000,
  number_not_literate_females = 100000
)

}
