% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGhanaCensus-package.R
\docType{package}
\name{rGhanaCensus-package}
\alias{rGhanaCensus}
\alias{rGhanaCensus-package}
\title{rGhanaCensus: 2021 Ghana Population and Housing Census Results as Data Frames}
\description{
Datasets from the 2021 Ghana Population and Housing Census Results. Users can access results as 'tidyverse' and 'sf'-Ready Data Frames. The data in this package is downloaded or scraped from pdf reports released by the Ghana Statistical Service website \url{https://census2021.statsghana.gov.gh/} .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ktemadarko/rGhanaCensus}
  \item \url{https://ktemadarko.github.io/rGhanaCensus/}
  \item Report bugs at \url{https://github.com/ktemadarko/rGhanaCensus/issues}
}

}
\author{
\strong{Maintainer}: Ama Owusu-Darko \email{aowusuda@asu.edu} (\href{https://orcid.org/0000-0002-7419-0324}{ORCID}) [copyright holder]

}
\keyword{internal}
