% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGhanaCensus_parquet_format.R
\name{rGhanaCensus_parquet_format}
\alias{rGhanaCensus_parquet_format}
\title{Get path to census data in parquet format}
\usage{
rGhanaCensus_parquet_format(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
rGhanaCensus stores the ten percent sample of the 2021 Ghana Population and Housing Census data in the parquet format in its
\code{inst/extdata} folder. This function make them easy to access.
}
\examples{
rGhanaCensus_parquet_format()
rGhanaCensus_parquet_format("pq")
}
